change to double instead of int to not lose precision

+++ main/model/color.h
@@ -17,7 +17,7 @@
      * If number is greater than range2, then the lightest shade of color
      * is chosen.
      */
-    void scale_color( double value, int maxVal, int minVal, int x, int y);
+    void scale_color( double value, double maxVal, double minVal, int x, int y);
 
     void update_color();
